feat: add modals folder with Order model - update components using order related functions - also add numbers utils
