+++ packages/app/app/models/order/types.ts
@@ -1,0 +1,25 @@
+export interface Token {
+  id: string;
+  decimals: number;
+  name: string;
+  symbol: string;
+}
+
+export interface Order {
+  id: string;
+  owner: string;
+  receiver: string;
+  amount: string;
+  sellToken: Token;
+  buyToken: Token;
+  createdAt: string;
+  startTime: number;
+  endTime: number;
+  orderSlots: string[];
+  cancelledAt: string | null;
+  interval: string;
+}
+
+export interface OrderProps {
+  order: Order;
+}
