+++ packages/app/app/stacks/components/StackedTokenLogoPair.tsx
@@ -1,6 +1,6 @@
-import { Order } from "@/app/stacks/page";
+import { OrderProps } from "@/app/models/order";
 
-export const StackedTokenLogoPair = ({ order }: { order: Order }) => (
+export const StackedTokenLogoPair = ({ order }: OrderProps) => (
   <div className="flex items-end">
     <div className="flex items-center justify-center w-10 h-10 text-[10px] border-2 rounded-full border-primary-400 bg-primary-100">
       {order.buyToken.symbol.substring(0, 4)}
