+++ packages/app/utils/numbers.ts
@@ -1,0 +1,3 @@
+// we'll probably use viem/ethers function when we connect to data
+export const convertedAmount = (amount: string | number, decimals: number) =>
+  Number(amount) / 10 ** decimals;
